Created separated MASTER and SLAVE Modes of operation
The only impacted function was the DaliRXDecoding() function, which was
added another attribute called mode. When in Slave Mode, after receiving
the first Byte it jumps to DATA state and  it receives the second byte.
If in Master Mode it receives one byte and it goes to the STOP State